backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.9.2
pydantic-settings==2.6.1
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0
redis==5.2.0
rq==1.16.2
boto3==1.35.66
faster-whisper==1.1.0
demucs==4.0.1
lyricsgenius==3.0.1
rapidfuzz==3.10.1
slowapi==0.1.9
python-multipart==0.0.17
